Fixed float in in-air tail tune

+++ src/main/flight/mixer_tricopter.h
@@ -167,7 +167,7 @@ typedef enum {
 } tailtuneMode_e;
 
 typedef struct servoAvgAngle_s {
-    uint32_t sum;
+    float sum;
     uint16_t numOf;
 } servoAvgAngle_t;
